ajout du token d'authentification et du mur d'authen pour protéger les routes

+++ userHandlers.js
@@ -83,6 +83,11 @@ const updateUser = (req, res) => {
   const { firstname, lastname, email, city, language, hashedPassword } =
     req.body;
 
+  // Vérifier si l'ID correspond à celui du payload du token
+  if (id !== req.payload.sub) {
+    return res.sendStatus(403); // Renvoi statut Forbidden
+  }
+
   if (!hashedPassword) {
     res.status(400).send("Password is required");
     return;
@@ -109,6 +114,11 @@ const updateUser = (req, res) => {
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
 
+  // Vérifier si l'ID correspond à celui du payload du token
+  if (id !== req.payload.sub) {
+    return res.sendStatus(403); // Renvoi statut Forbidden
+  }
+
   database
     .query("delete from users where id = ?", [id])
     .then(([result]) => {
@@ -124,10 +134,31 @@ const deleteUser = (req, res) => {
     });
 };
 
+const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
+  const { email } = req.body;
+
+  database
+    .query("select * from users where email = ?", [email])
+    .then(([users]) => {
+      if (users[0] != null) {
+        req.user = users[0];
+
+        next();
+      } else {
+        res.sendStatus(401);
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error retrieving data from database");
+    });
+};
+
 module.exports = {
   getUsers,
   getUserById,
   postUser,
   updateUser,
   deleteUser,
+  getUserByEmailWithPasswordAndPassToNext,
 };
